requirements.txt
fastapi
uvicorn[standard]
python-dotenv
xgboost
scikit-learn
shap
pandas
numpy
openpyxl
joblib
matplotlib
sse-starlette
google-generativeai
mlflow
